moving nndownload.py into its own module
